Clean up GROUPBY queries based on a precise semantics for GROUPBY.
1. Now, there is no operation following a GROUPBY unless it emits.
If it emits, the GROUPBY produces a stream, which can be fed to other
operators. If not, the GROUPBY produces a relation, which can't be
transformed further.
(Still confused about the semantics of high_e2e_latency and high_tail_latency)

2. Also, we don't need a spurious project after each GROUPBY (e.g.,
pkt_counts and byte_counts before this commit). GROUPBY is now defined
to return all fields in the GROUPBY clause and all state in the
agg function (cf. comment at the bottom of perf_query.g4)